Do not fail on local shared objects
If the binary depends on a library whose soname is a path, do not ignore
it, use the soname as a path instead.

We used to ignore it because such libraries because this is how the
ld-linux.so appear in ldd output, but this can actually happen for
libraries one wants to ship. The case of ld-linux.so is taken care of when
we traverse the tree since they are in the default blacklist.

Fixes #7